requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0
opencv-python-headless>=4.8.0
facenet-pytorch>=2.5.0
insightface>=0.7.0
onnxruntime-gpu>=1.16.0
ftfy
regex
open_clip_torch>=2.20.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0
pyyaml>=6.0
tqdm>=4.65.0
fvcore>=0.1.5

# --- Dev / testing ---
pytest>=7.4.0
pytest-cov>=4.1.0
